refactor(web): complete distance cap/submit branches; extract dialog filters
Address Copilot review on #485:
- add the `distance` measurement branch to exerciseOverCapKind and
  canSubmitExercise so the dead-but-maintained distance path stays consistent
  with buildExerciseResult (behaviour-neutral today — no catalog exercise uses
  `measurement: 'distance'`); covered by submit.spec
- move the pushup autocomplete filters into training-entry-dialog.filters.ts so
  submit.ts stays ≤250 LOC after the additions

All dialog prod files ≤250; 100 dialog tests pass.

Refs #456
